app/requirements.txt
asgiref==3.8.1
babel==2.17.0
Django==5.2.1
django-environ==0.12.0
django-filter==25.1
django-money==3.5.4
django-requestlogs==0.8.3
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
py-moneyed==3.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.5
pytz==2025.2
PyYAML==6.0.2
setuptools==80.8.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
